Move run_subprocess_side_effect to testing utils
