+++ convert2rhel/unit_tests/__init__.py
@@ -233,3 +233,33 @@ class GetFileContentMocked(MockFunction):
         self.called += 1
         self.as_list = as_list
         return [x.strip() for x in self.data] if self.as_list else self.data
+
+
+def run_subprocess_side_effect(*stubs):
+    """Side effect function for utils.run_subprocess.
+    :type stubs: Tuple[Tuple[command, ...], Tuple[command_stdout, exit_code]]
+
+    Allows you to parametrize the mocking by providing list of stubs
+
+    if run_subprocess called with args, which are not specified in
+    stubs, then no mocking happening and a real subprocess call will be made.
+
+    Example:
+    >>> run_subprocess_mock = mock.Mock(
+    >>>     side_effect=run_subprocess_side_effect(
+    >>>         (("uname",), ("5.8.0-7642-generic\n", 0)),
+    >>>         (("repoquery", "-f"), (REPOQUERY_F_STUB_GOOD, 0)),
+    >>>         (("repoquery", "-l"), (REPOQUERY_L_STUB_GOOD, 0)),
+    >>>     )
+    >>> )
+
+    """
+
+    def factory(*args, **kwargs):
+        for kws, result in stubs:
+            if all(kw in args[0].split() for kw in kws):
+                return result
+        else:
+            return run_subprocess(*args, **kwargs)
+
+    return factory
